BREAKING: use relative imports in costs module
