+++ skchange/costs/_l2_cost.py
@@ -1,47 +1,15 @@
-"""Univariate Gaussian likelihood cost."""
-
-__author__ = ["Tveten"]
+"""L2 cost."""
 
 import numpy as np
 
-from skchange.costs.base import BaseCost
-from skchange.costs.utils import MeanType, VarType, check_mean, check_var
-from skchange.utils.numba import njit
-from skchange.utils.numba.general import truncate_below
-from skchange.utils.numba.stats import col_cumsum
+from ..costs._utils import MeanType, check_mean
+from ..utils.numba import njit
+from ..utils.numba.stats import col_cumsum
+from .base import BaseCost
 
 
 @njit
-def var_from_sums(
-    sums: np.ndarray, sums2: np.ndarray, starts: np.ndarray, ends: np.ndarray
-):
-    """Calculate variance from precomputed sums.
-
-    Parameters
-    ----------
-    sums : np.ndarray
-        Cumulative sums of ``X``, where the first row are ``0``s.
-    sums2 : np.ndarray
-        Cumulative sums of ``X**2``, where the first row are ``0``s.
-    starts : np.ndarray
-        Start indices in the original data ``X``.
-    ends : np.ndarray
-        End indices in the original data ``X``.
-
-    Returns
-    -------
-    var : float
-        Variance of ``X[i:j]``.
-    """
-    n = (ends - starts).reshape(-1, 1)
-    partial_sums = sums[ends] - sums[starts]
-    partial_sums2 = sums2[ends] - sums2[starts]
-    var = partial_sums2 / n - (partial_sums / n) ** 2
-    return truncate_below(var, 1e-16)  # standard deviation lower bound of 1e-8
-
-
-@njit
-def gaussian_var_cost_optim(
+def l2_cost_optim(
     starts: np.ndarray,
     ends: np.ndarray,
     sums: np.ndarray,
@@ -56,31 +24,32 @@ def gaussian_var_cost_optim(
     ends : np.ndarray
         End indices of the segments.
     sums : np.ndarray
-        Cumulative sum of the input data, with a row of ``0``-entries as the first row.
+        Cumulative sum of the input data, with a row of 0-entries as the first row.
     sums2 : np.ndarray
-        Cumulative sum of the squared input data, with a row of ``0``-entries as the
-        first row.
+        Cumulative sum of the squared input data, with a row of 0-entries as the first
+        row.
 
     Returns
     -------
     costs : np.ndarray
         A 2D array of costs. One row for each interval. The number of columns
-        is equal to the number of columns in the input data.
+        is equal to the number of columns in the input data, where each column
+        represents the univariate cost for the corresponding input data column.
     """
+    partial_sums = sums[ends] - sums[starts]
+    partial_sums2 = sums2[ends] - sums2[starts]
     n = (ends - starts).reshape(-1, 1)
-    var = var_from_sums(sums, sums2, starts, ends)
-    log_likelihood = -n * np.log(2 * np.pi * var) - n
-    return -log_likelihood
+    costs = partial_sums2 - partial_sums**2 / n
+    return costs
 
 
 @njit
-def gaussian_var_cost_fixed(
+def l2_cost_fixed(
     starts: np.ndarray,
     ends: np.ndarray,
     sums: np.ndarray,
     sums2: np.ndarray,
     mean: np.ndarray,
-    var: np.ndarray,
 ) -> np.ndarray:
     """Calculate the L2 cost for a fixed constant mean for each segment.
 
@@ -91,38 +60,35 @@ def gaussian_var_cost_fixed(
     ends : np.ndarray
         End indices of the segments.
     sums : np.ndarray
-        Cumulative sum of the input data, with a row of ``0``-entries as the first row.
+        Cumulative sum of the input data, with a row of 0-entries as the first row.
     sums2 : np.ndarray
-        Cumulative sum of the squared input data, with a row of ``0``-entries as the
-        first row.
+        Cumulative sum of the squared input data, with a row of 0-entries as the first
+        row.
     mean : np.ndarray
         Fixed mean for the cost calculation.
-    var : np.ndarray
-        Fixed variance for the cost calculation.
 
     Returns
     -------
     costs : np.ndarray
         A 2D array of costs. One row for each interval. The number of columns
-        is equal to the number of columns in the input data.
+        is equal to the number of columns in the input data, where each column
+        represents the univariate cost for the corresponding input data column.
     """
-    n = (ends - starts).reshape(-1, 1)
     partial_sums = sums[ends] - sums[starts]
     partial_sums2 = sums2[ends] - sums2[starts]
+    n = (ends - starts).reshape(-1, 1)
+    costs = partial_sums2 - 2 * mean * partial_sums + n * mean**2
+    return costs
 
-    quadratic_form = partial_sums2 - 2 * mean * partial_sums + n * mean**2
-    log_likelihood = -n * np.log(2 * np.pi * var) - quadratic_form / var
-    return -log_likelihood
 
-
-class GaussianCost(BaseCost):
-    """Univariate Gaussian likelihood cost.
+class L2Cost(BaseCost):
+    """L2 cost of a constant mean.
 
     Parameters
     ----------
-    param : 2-tuple of float or np.ndarray, or None (default=None)
-        Fixed mean(s) and variance(s) for the cost calculation.
-        If ``None``, the maximum likelihood estimates are used.
+    param : float or array-like, optional (default=None)
+        Fixed mean for the cost calculation. If ``None``, the optimal mean is
+        calculated.
     """
 
     _tags = {
@@ -132,18 +98,16 @@ class GaussianCost(BaseCost):
 
     supports_fixed_params = True
 
-    def __init__(self, param: tuple[MeanType, VarType] | None = None):
+    def __init__(self, param: MeanType | None = None):
         super().__init__(param)
 
-    def _check_fixed_param(
-        self, param: tuple[MeanType, VarType], X: np.ndarray
-    ) -> np.ndarray:
+    def _check_fixed_param(self, param: MeanType, X: np.ndarray) -> np.ndarray:
         """Check if the fixed mean parameter is valid.
 
         Parameters
         ----------
-        param : 2-tuple of float or np.ndarray, or None (default=None)
-            Fixed mean(s) and variance(s) for the cost calculation.
+        param : float or array-like
+            The input parameter to check.
         X : np.ndarray
             Input data.
 
@@ -152,28 +116,7 @@ class GaussianCost(BaseCost):
         mean : np.ndarray
             Fixed mean for the cost calculation.
         """
-        mean, var = param
-        mean = check_mean(mean, X)
-        var = check_var(var, X)
-        return mean, var
-
-    @property
-    def min_size(self) -> int:
-        """Minimum size of the interval to evaluate.
-
-        The size of each interval is defined as ``intervals[i, 1] - intervals[i, 0]``.
-        """
-        return 2
-
-    def get_param_size(self, p: int) -> int:
-        """Get the number of parameters in the cost function.
-
-        Parameters
-        ----------
-        p : int
-            Number of variables in the data.
-        """
-        return 2 * p
+        return check_mean(param, X)
 
     def _fit(self, X: np.ndarray, y=None):
         """Fit the cost.
@@ -187,10 +130,11 @@ class GaussianCost(BaseCost):
         y: None
             Ignored. Included for API consistency by convention.
         """
-        self._param = self._check_param(self.param, X)
+        self._mean = self._check_param(self.param, X)
 
         self._sums = col_cumsum(X, init_zero=True)
         self._sums2 = col_cumsum(X**2, init_zero=True)
+
         return self
 
     def _evaluate_optim_param(self, starts: np.ndarray, ends: np.ndarray) -> np.ndarray:
@@ -210,7 +154,7 @@ class GaussianCost(BaseCost):
             is equal to the number of columns in the input data, where each column
             represents the univariate cost for the corresponding input data column.
         """
-        return gaussian_var_cost_optim(starts, ends, self._sums, self._sums2)
+        return l2_cost_optim(starts, ends, self._sums, self._sums2)
 
     def _evaluate_fixed_param(self, starts, ends):
         """Evaluate the cost for the fixed parameter.
@@ -229,8 +173,7 @@ class GaussianCost(BaseCost):
             is equal to the number of columns in the input data, where each column
             represents the univariate cost for the corresponding input data column.
         """
-        mean, var = self._param
-        return gaussian_var_cost_fixed(starts, ends, self._sums, self._sums2, mean, var)
+        return l2_cost_fixed(starts, ends, self._sums, self._sums2, self._mean)
 
     @classmethod
     def get_test_params(cls, parameter_set="default"):
@@ -240,7 +183,7 @@ class GaussianCost(BaseCost):
         ----------
         parameter_set : str, default="default"
             Name of the set of test parameters to return, for use in tests. If no
-            special parameters are defined for a value, will return "default" set.
+            special parameters are defined for a value, will return `"default"` set.
             There are currently no reserved values for interval evaluators.
 
         Returns
@@ -253,7 +196,6 @@ class GaussianCost(BaseCost):
         """
         params = [
             {"param": None},
-            {"param": (0.0, 1.0)},
-            {"param": (np.zeros(1), np.ones(1))},
+            {"param": 0.0},
         ]
         return params
